Move `channelBroadcaster` into internal package
