+++ lib/tbot/service_ssh_host_output.go
@@ -39,6 +39,7 @@ import (
 	"github.com/gravitational/teleport/lib/tbot/client"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/tbot/identity"
+	"github.com/gravitational/teleport/lib/tbot/internal"
 	"github.com/gravitational/teleport/lib/tbot/readyz"
 )
 
@@ -49,7 +50,7 @@ type SSHHostOutputService struct {
 	cfg                *config.SSHHostOutput
 	getBotIdentity     getBotIdentityFn
 	log                *slog.Logger
-	reloadBroadcaster  *channelBroadcaster
+	reloadBroadcaster  *internal.ChannelBroadcaster
 	statusReporter     readyz.Reporter
 	identityGenerator  *identity.Generator
 	clientBuilder      *client.Builder
@@ -67,7 +68,7 @@ func (s *SSHHostOutputService) OneShot(ctx context.Context) error {
 }
 
 func (s *SSHHostOutputService) Run(ctx context.Context) error {
-	reloadCh, unsubscribe := s.reloadBroadcaster.subscribe()
+	reloadCh, unsubscribe := s.reloadBroadcaster.Subscribe()
 	defer unsubscribe()
 
 	err := runOnInterval(ctx, runOnIntervalConfig{
